[FLINK-37818] Move KafkaCommitter to internal
There is no leaking of the specific class in any of the KafkaSink(Builder) signatures.

(cherry picked from commit a18d7afa17027e40dd35cf1586edd9389ee6cecc)
